feat: implement recording improvements and organized database system
Major Improvements:
- Fixed Filesize type from int to int64 (supports files >2GB)
- Added stream end detection (10s quick retry)
- Added MP4 init segment immediate sync
- Added periodic file sync (every 10 segments)
- Added exponential backoff for Cloudflare blocks
- Added disk space pre-flight check

Database System:
- Organized by channel name (one folder per channel)
- Organized by date (one JSON file per date)
- Auto-commit after uploads
- Auto-commit on emergency/timeout
- Auto-commit at workflow completion
- Retry logic with 3 attempts

Technical Changes:
- channel/channel.go: Changed Filesize to int64, added segmentCount
- channel/channel_record.go: All improvements implemented
- channel/channel_file.go: Fixed type conversions
- internal/internal.go: FormatFilesize now accepts int64
- internal/internal_err.go: Added ErrStreamEnded, ErrDiskSpaceCritical
- manager/manager.go: Added CheckDiskSpace() method
- server/manager.go: Added CheckDiskSpace() to interface
- .github/workflows/continuous-runner.yml: Database auto-commit system

Documentation:
- docs/RECORDING_ARCHITECTURE_COMPARISON.md
- docs/IMPROVEMENTS_ANALYSIS.md
- docs/IMPROVEMENTS_IMPLEMENTED.md
- docs/PROPER_IMPLEMENTATION_SUMMARY.md
- docs/DATABASE_AUTO_COMMIT.md
- docs/DATABASE_STRUCTURE.md
- database/README.md

All changes compile successfully and maintain 100% backward compatibility.

+++ channel/channel_file.go
@@ -241,7 +241,7 @@ func (ch *Channel) ShouldSwitchFile() bool {
 }
 
 func (ch *Channel) shouldSwitchFileLocked() bool {
-	maxFilesizeBytes := ch.Config.MaxFilesize * 1024 * 1024
+	maxFilesizeBytes := int64(ch.Config.MaxFilesize) * 1024 * 1024
 	maxDurationSeconds := ch.Config.MaxDuration * 60
 
 	return (ch.Duration >= float64(maxDurationSeconds) && ch.Config.MaxDuration > 0) ||
